Remove unused `ostruct`
Since Ruby 3.3.5, Bundler shows the following warning when using
`ostruct`.

```
ostruct was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.
```
Ref: https://bugs.ruby-lang.org/issues/20309

As the message described, we need to add `ostruct` to the gemspec
to keep using it. But, it seems that `diff-lcs` doesn't use it.
So I just remove `require of it.