Fix hunk/ldiff unified diff against empty files
The ldiff bin (and it's underlying implementation) behaves weirdly when
given an empty file as one of its operand (old or new file).

This is particularly painful when doing diff for new or deleted files,
especially as one could think it is a simple case as the whole content
should be considered either added or removed.

To demonstrate the issue, I'm adding test cases to demonstrate
how ldiff behaves when a 4-lines file is compared to empty, using all
supported formats and in both directions.

The expected result has been collected using GNU diff 3.10.