Extract ldiff display logic to reuse it as lib
The ldiff "display" logic was too tightly coupled with the command line
program. Let's just extract it in a distinct class method accepting all
the necessary parameters.

One can now use
  Diff::LCS::Ldiff.diff?(
    Diff::LCS::Ldiff::InputInfo.new('file1'),
    Diff::LCS::Ldiff::InputInfo.new('file2'),
    :unified, # or :ed or :context or :report
    $stdout, # or a ref to [] or to a StringIO or a File
    binary: nil, # or true or false if you want to be explicit
    lines: 3) # to customize the number of context lines

The result will be `true` in case of diff and `false` otherwise.
The output if any, will be appended to the output parameter.