Fix ldiff binary files comparison and output
Before this change, the comparison was reversed and diff were not
detected.

This commit also change the output when a diff is detected to look
like the GNU diff output.

Also fix the support of all other format than :report, by quickly
halting the diff process if there is a binary diff.