Fix hunk.diff "No newline at end of file" handling
The unified format was almost correct, the only failing cases were
diff against empty files (old or new).
On the other hand the ed, context & old_diff formats were just ignoring
this case.

The basic principle used is to display the message only:
- on the last hunk i.e. diff(format, last=true)
- if the data is not empty

The ed format is not perfectly handled, it differs from the GNU diff
stderr output
- GNU diff: "diff <path_to_the_file>: No newline at end of file\n\n"
- ldiff: "<(old|new)_file>: No newline at end of file\n"

Rationale are:
1. hunk system is not "filename" aware
2. since output is different (because of above point) let's not use
   two newline