Fix hunk ranges computation
All the supported format of the hunk system are computing their ranges
using the @start_old, @start_end, @start_new, @end_new variables that
are defined at initialization.

The check to ensure the context won't go beyond the old_size was not
correct as it used a strict "greater than" comparison.
As ruby array's indexes are 0 based, is important that the index does
not equal the size of the array.

The upper boundary check was meant to fit the context to the end of the
data, but the computation of the "add_end" value was also incorrect,
very likely related to the previous error.

Also certaintly related to previous errors, the "add_end" correction
was reset to "@max_diff_size" if "add_end >= old_size".
However due this should never happen, it would otherwise means that
one of the index or size was negative.

Finally, the context_range & unified_range method were trying to fix
the indexes based on some heuristic (mostly the "last" parameter).
With the previous fixes, this is not necessary anymore, the indexes
are now correct.