chore: Drop ed & reverse_ed formats
These formats are super old and very unlikely used. Their implementation
was broken for a while and no-one reported it.

The GNU doc say that ed format is "almost obsolete". [1]
The reverse_ed (aka forward-ed) is described as "not very useful". [2]

[1]: https://www.gnu.org/software/diffutils/manual/html_node/ed-Scripts.html
[2]: https://www.gnu.org/software/diffutils/manual/html_node/Forward-ed.html

Resolves: #108
Breaking-Change: The ed and reverse_ed output formats are no longer supported in ldiff.
Signed-off-by: Austin Ziegler <austin@zieglers.ca>