Replace usage of `cgi` gem for Ruby 3.5
In Ruby 3.5 most of the cgi gem will be removed.
Only the various escape/unescape methods will be retained by default.

But, it is a bit cumbersome to find a uniform solution that works on
all supported rubies, so use the same functionality provided
via `erb` instead.

I've confirmed this works down to Ruby 2.0

https://bugs.ruby-lang.org/issues/21258